Moved call to adapter package for model view adapter
